add stack trace to Gate watcher
