+++ public/app.js
@@ -50567,7 +50567,25 @@ var render = function() {
                   ]
                 )
               ])
-            ])
+            ]),
+            _vm._v(" "),
+            slotProps.entry.content.file
+              ? _c("tr", [
+                  _c("td", { staticClass: "table-fit font-weight-bold" }, [
+                    _vm._v("Location")
+                  ]),
+                  _vm._v(" "),
+                  _c("td", [
+                    _vm._v(
+                      "\n                " +
+                        _vm._s(slotProps.entry.content.file) +
+                        ":" +
+                        _vm._s(slotProps.entry.content.line) +
+                        "\n            "
+                    )
+                  ])
+                ])
+              : _vm._e()
           ]
         }
       },
