+++ public/mix-manifest.json
@@ -1,5 +1,5 @@
 {
-    "/app.js": "/app.js?id=8b839cf52d8d2895ec73",
+    "/app.js": "/app.js?id=cfc93f63ab192bcfa68d",
     "/app.css": "/app.css?id=7c1ff2a14db6cf79b2c3",
     "/app-dark.css": "/app-dark.css?id=3499432c9dbc93b0f541"
 }
