+++ resources/js/screens/gates/preview.vue
@@ -34,6 +34,13 @@
                     </span>
                 </td>
             </tr>
+
+            <tr  v-if="slotProps.entry.content.file">
+                <td class="table-fit font-weight-bold">Location</td>
+                <td>
+                    {{slotProps.entry.content.file}}:{{slotProps.entry.content.line}}
+                </td>
+            </tr>
         </template>
 
         <div slot="after-attributes-card" slot-scope="slotProps">
