+++ src/Watchers/QueryWatcher.php
@@ -2,13 +2,14 @@
 
 namespace Laravel\Telescope\Watchers;
 
-use Illuminate\Support\Str;
 use Laravel\Telescope\Telescope;
 use Laravel\Telescope\IncomingEntry;
 use Illuminate\Database\Events\QueryExecuted;
 
 class QueryWatcher extends Watcher
 {
+    use FetchesStackTrace;
+
     /**
      * Register the watcher.
      *
@@ -67,37 +68,5 @@ class QueryWatcher extends Watcher
     protected function formatBindings($event)
     {
         return $event->connection->prepareBindings($event->bindings);
-    }
-
-    /**
-     * Find the first frame in the stack trace outside of Telescope/Laravel.
-     *
-     * @return array
-     */
-    protected function getCallerFromStackTrace()
-    {
-        $trace = collect(debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS))->forget(0);
-
-        return $trace->first(function ($frame) {
-            if (! isset($frame['file'])) {
-                return false;
-            }
-
-            return ! Str::contains($frame['file'],
-                base_path('vendor'.DIRECTORY_SEPARATOR.$this->ignoredVendorPath())
-            );
-        });
-    }
-
-    /**
-     * Choose the frame outside of either Telescope/Laravel or all packages.
-     *
-     * @return string|null
-     */
-    protected function ignoredVendorPath()
-    {
-        if (! ($this->options['ignore_packages'] ?? true)) {
-            return 'laravel';
-        }
     }
 }
